move cache classes to use case
