+++ src/main/kotlin/org/bibletranslationtools/fetcher/usecase/cache/AvailabilityCache.kt
@@ -1,4 +1,4 @@
-package org.bibletranslationtools.fetcher.data.cache
+package org.bibletranslationtools.fetcher.usecase.cache
 
 data class AvailabilityCache(
     var languages: List<LanguageCache>
